IP whitelist per user

+++ store/sqlstore/whitelist_store.go
@@ -6,6 +6,7 @@ package sqlstore
 import (
 	"github.com/pkg/errors"
 
+	sq "github.com/Masterminds/squirrel"
 	"github.com/zacmm/zacmm-server/einterfaces"
 	"github.com/zacmm/zacmm-server/model"
 	"github.com/zacmm/zacmm-server/store"
@@ -22,7 +23,8 @@ func newSqlWhitelistStore(sqlSupplier *SqlSupplier) store.WhitelistStore {
 	}
 
 	for _, db := range sqlSupplier.GetAllConns() {
-		table := db.AddTableWithName(model.WhitelistItem{}, "Whitelist").SetKeys(false, "IP")
+		table := db.AddTableWithName(model.WhitelistItem{}, "Whitelist").SetKeys(false, "UserId", "IP")
+		table.ColMap("UserId").SetMaxSize(26)
 		table.ColMap("IP").SetMaxSize(39)
 	}
 
@@ -34,32 +36,37 @@ func (s SqlWhitelistStore) createIndexesIfNotExists() {
 
 func (s SqlWhitelistStore) Add(whitelistItem *model.WhitelistItem) error {
 
+	if len(whitelistItem.UserId) == 0 {
+		return store.NewErrInvalidInput("whitelist item", "user id", whitelistItem.UserId)
+	}
+
 	if len(whitelistItem.IP) == 0 {
 		return store.NewErrInvalidInput("whitelist item", "ip", whitelistItem.IP)
 	}
 
 	if err := s.GetMaster().Insert(whitelistItem); err != nil {
-		return errors.Wrapf(err, "failed to save whitelist item with ip=%s", whitelistItem.IP)
+		return errors.Wrapf(err, "failed to save whitelist item with user_id=%s and ip=%s", whitelistItem.UserId, whitelistItem.IP)
 	}
 
 	return nil
 }
 
 func (s SqlWhitelistStore) Delete(whitelistItem *model.WhitelistItem) error {
-	_, err := s.GetMaster().Exec("DELETE FROM Whitelist WHERE IP = :IP", map[string]interface{}{"IP": whitelistItem.IP})
+	_, err := s.GetMaster().Exec("DELETE FROM Whitelist WHERE UserId = :UserId AND IP = :IP", map[string]interface{}{"UserId": whitelistItem.UserId, "IP": whitelistItem.IP})
 	if err != nil {
-		return errors.Wrapf(err, "failed to delete from Whitelist with ip=%s", whitelistItem.IP)
+		return errors.Wrapf(err, "failed to delete from Whitelist with user id=%s and ip=%s", whitelistItem.UserId, whitelistItem.IP)
 	}
 
 	return nil
 }
 
-func (s SqlWhitelistStore) Get() ([]string, error) {
+func (s SqlWhitelistStore) GetByUserId(userId string) ([]string, error) {
 	var ips []string
 
 	query := s.getQueryBuilder().
 		Select("IP").
-		From("Whitelist")
+		From("Whitelist").
+		Where(sq.Eq{"UserId": userId})
 
 	queryString, args, err := query.ToSql()
 	if err != nil {
